Various small changes
iris-coordination-arbiter:
  * Change priv->active from glong to gint, and always access it from inside
    arbiter's mutex rather than with g_atomics functions
  * Fix initial state

iris-port:
  * Remove some dead code

iris-receiver:
  * iris_receiver_deliver_real(): rearrange a tiny bit to improve a atomicity
  * iris_receiver_worker_cb(): rename to iris_receiver_worker_destroy() and
    rework calling slightly (since scheduler execution notification is dead)

tests/receiver-scheduler-1:
  * Add g_test repetition framework; could be useful in other tests too
  * Repeat live tests a lot & alter slightly, to trigger a race condition
    (still unfixed) that breaks iris_port_flush() and causes the port queue to
    freeze.

+++ iris/iris-coordination-arbiter-private.h
@@ -52,7 +52,7 @@ struct _IrisCoordinationArbiterPrivate
 	IrisReceiver    *teardown;
 	GStaticRecMutex  mutex;
 	guint            flags;
-	glong            active;
+	gint             active;
 };
 
 #endif /* __IRIS_COORDINATION_ARBITER_PRIVATE_H__ */
